refactor!: remove json.h because there exists way better solutions
BREAKING CHANGE: A good alternative is yaml-cpp and that's what I'm going to be using. No need to wrap the API because it's very clear on it's own (i.e. YAML::LoadFile(...)). It has a concise way of defining deserializers, hex values are handles properly by default (so no need for parseHex...), and since json is a subset of yaml, it's a good library for reading json. Use yaml-cpp, for real, yo